refactor: modularise caseplan
Updated mock paths in test_caseplan.py to reflect new module structure by adding specific submodule references (budget_assessor and plan_generator) to the patched imports. This ensures tests correctly target the updated code organization.
