+++ litassist/commands/caseplan/__init__.py
@@ -1,0 +1,4 @@
+"""Case-specific litigation workflow planning command package."""
+from .core import caseplan
+
+__all__ = ["caseplan"]
